problems on statistics
2.​ Problem Statement

Task 1:
1. You survey households in your area to find the average rent they are paying. Find the
standard deviation from the following data:
$1550, $1700, $900, $850, $1000, $950.

2. Find the variance for the following set of data representing trees in California (heights in
feet):
3, 21, 98, 203, 17, 9

3. In a class on 100 students, 80 students passed in all subjects, 10 failed in one subject, 7 failed
in two subjects and 3 failed in three subjects. Find the probability distribution of the variable for
number of subjects a student from the given class has failed in.

Task 2:

1. A test is conducted which is consisting of 20 MCQs (multiple choices questions) with every
MCQ having its four options out of which only one is correct. Determine the probability that a
person undertaking that test has answered exactly 5 questions wrong.

2. A die marked A to E is rolled 50 times. Find the probability of getting a “D” exactly 5 times.

3. Two balls are drawn at random in succession without replacement from an urn containing 4
red balls and 6 black balls.

Find the probabilities of all the possible outcomes.